+++ src/math/Matrix4f.cpp
@@ -62,14 +62,14 @@ Matrix4f Matrix4f::Rotation_Matrix(float x, float y, float z)
 	y = to_radians(y);
 	z = to_radians(z);
 
-	rx.set(0, 0, std::cosf(z)), rx.set(0, 1, -std::sinf(z));
-	rx.set(1, 0, std::sinf(z)), rx.set(1, 1, std::cosf(z));
+	rx.set(0, 0, std::cos(z)), rx.set(0, 1, -std::sin(z));
+	rx.set(1, 0, std::sin(z)), rx.set(1, 1, std::cos(z));
 
-	ry.set(1, 1, std::cosf(x)), ry.set(1, 2, -std::sinf(x));
-	ry.set(2, 1, std::sinf(x)), ry.set(2, 2, std::cosf(x));
+	ry.set(1, 1, std::cos(x)), ry.set(1, 2, -std::sin(x));
+	ry.set(2, 1, std::sin(x)), ry.set(2, 2, std::cos(x));
 
-	rz.set(0, 0, std::cosf(y)), rz.set(0, 2, -std::sinf(y));
-	rz.set(2, 0, std::sinf(y)), rz.set(2, 2, std::cosf(y));
+	rz.set(0, 0, std::cos(y)), rz.set(0, 2, -std::sin(y));
+	rz.set(2, 0, std::sin(y)), rz.set(2, 2, std::cos(y));
 
 	return rz * (ry * rx);
 }
